Move notification dispatch logic into a JS helper
